// src/types/Challenges.ts
export interface Challenges {
  _id: string
  gameId: string
  playerGuestId: string
  playerHostId: string
  typeMatch: string
  value: number
  durationInvite: number
  message: string
  stage: string
  createdAt: string
  updatedAt: string
  __v: number
  V: number
  JR: number
  D: number
  status: string
  games: any
}